requirements/base.txt
Django>=1.10
django-environ>=0.4.0
django-braces>=1.8.1
# python-ldap>=2.4.25 (needs to be installed through wheel)